requirements.txt
rembg
pillow
requests
numpy
